requirements.txt
pywin32==301
loguru==0.6.0
pyautogui==0.9.53
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
numpy==1.21.6
PyYAML==6.0
aircv==1.4.6
